Replace references to "Discrete" executions with simply Executions; deprecate `GoodJob::DiscreteExecution`
