+++ spec/app/jobs/example_job_spec.rb
@@ -27,8 +27,8 @@ describe ExampleJob do
         Timecop.return
 
         good_job = GoodJob::Job.find_by(active_job_id: active_job.job_id)
-        expect(good_job.discrete_executions.count).to eq 2
-        expect(good_job.discrete_executions.last.error).to be_nil
+        expect(good_job.executions.count).to eq 2
+        expect(good_job.executions.last.error).to be_nil
       end
     end
 
@@ -43,8 +43,8 @@ describe ExampleJob do
 
         good_job = GoodJob::Job.find_by(active_job_id: active_job.job_id)
 
-        expect(good_job.discrete_executions.count).to eq 6
-        expect(good_job.discrete_executions.last.error).to be_nil
+        expect(good_job.executions.count).to eq 6
+        expect(good_job.executions.last.error).to be_nil
       end
     end
 
@@ -58,11 +58,11 @@ describe ExampleJob do
         Timecop.return
 
         good_job = GoodJob::Job.find_by(active_job_id: active_job.job_id)
-        expect(good_job.discrete_executions.count).to eq 3
-        discrete_execution = good_job.discrete_executions.last
-        expect(discrete_execution.error).to be_present
-        expect(discrete_execution.error_backtrace.count).to be > 100
-        expect(discrete_execution.filtered_error_backtrace).to eq(["app/jobs/example_job.rb:41:in `perform'"])
+        expect(good_job.executions.count).to eq 3
+        execution = good_job.executions.last
+        expect(execution.error).to be_present
+        expect(execution.error_backtrace.count).to be > 100
+        expect(execution.filtered_error_backtrace).to eq(["app/jobs/example_job.rb:41:in `perform'"])
       end
     end
