+++ spec/integration/batch_spec.rb
@@ -216,7 +216,7 @@ RSpec.describe 'Batches' do
       GoodJob.perform_inline
 
       expect(GoodJob::Job.count).to eq 3
-      expect(GoodJob::DiscreteExecution.count).to eq 5
+      expect(GoodJob::Execution.count).to eq 5
       expect(GoodJob::Job.where(batch_id: batch.id).count).to eq 1
       expect(GoodJob::Job.where(batch_callback_id: batch.id).count).to eq 2
 
@@ -231,7 +231,7 @@ RSpec.describe 'Batches' do
       GoodJob.perform_inline
 
       expect(GoodJob::Job.count).to eq 3
-      expect(GoodJob::DiscreteExecution.count).to eq 5
+      expect(GoodJob::Execution.count).to eq 5
       expect(GoodJob::Job.where(batch_id: batch.id).count).to eq 1
       expect(GoodJob::Job.where(batch_callback_id: batch.id).count).to eq 2
 
@@ -315,7 +315,7 @@ RSpec.describe 'Batches' do
       expect(job.performed_at).to be_present
       expect(job.finished_at).to be_present
 
-      job_executions = job.discrete_executions.order(:created_at).to_a
+      job_executions = job.executions.order(:created_at).to_a
       expect(job_executions.size).to eq 2
       expect(job_executions.first.status).to eq :discarded
       expect(job_executions.last.status).to eq :succeeded
